Mess with route heirarchy
- Refactors common layout style and logic
- Creates (app) layout group
- Adds `/embed/` route
