+++ src/routes/(app)/documents/+page.ts
