+++ src/routes/(app)/documents/[id]-[slug]/+layout.svelte
