+++ src/routes/(app)/documents/[id]-[slug]/+page.server.ts
