+++ src/routes/(app)/documents/[id]-[slug]/+page.ts
