+++ src/routes/(app)/documents/projects/+layout.ts
