+++ src/routes/(app)/documents/projects/[id]-[slug]/+page.server.ts
