+++ src/routes/(app)/documents/projects/[id]-[slug]/+page.ts
