+++ src/routes/(app)/documents/sidebar/AddOns.svelte
