--- src/routes/(embed)/+layout.svelte
@@ -1,6 +1,0 @@
-<script>
-  import "@/style/variables.css";
-  import "@/style/global.css";
-</script>
-
-<slot />
