+++ src/routes/embed/documents/[id]/+layout.ts
@@ -23,6 +23,8 @@ export async function load({ fetch, url, params, depends }) {
 
   depends(`document:${document.id}`);
 
+  console.log(document);
+
   let mode: ViewerMode =
     (url.searchParams.get("mode") as ViewerMode) ?? "document";
