+++ src/routes/embed/documents/[id]/+page.svelte
