+++ src/routes/embed/documents/[id]/+page.ts
