+++ src/routes/embed/documents/[id]/annotations/[note_id]/+page.ts
