+++ src/routes/embed/documents/[id]/pages/[page]/+page.ts
