+++ src/routes/embed/documents/[id]/pages/[page]/Note.svelte
