+++ src/routes/embed/projects/[project_id]/+page.svelte
